Move around Solr.
